+++ test/jest/__mock__/stripesInventoryComponents.mock.js
@@ -12,7 +12,11 @@ jest.mock('@folio/stripes-inventory-components', () => ({
     isLoading: false,
   }),
   useCommonData: jest.fn().mockReturnValue({
-    commonData: [],
+    commonData: {
+      displaySettings: {
+        defaultSort: 'title',
+      },
+    },
     isCommonDataLoading: false,
   }),
   InstanceFilters: jest.fn(() => <div>InstanceFilters</div>),